requirements.txt
python-dotenv==0.21.0
aiogram==2.23.1
